Adds ScopedTestKit
To get around Akka TestKit incompatibilities, we create a scoped trait that requires itself to be mixed into a TestKitBase, which in turn requires an implicit ActorSystem. We also include some implicit conversions from the two libraries respective Duration types to one another, and define a shutdown call that gets executed in the after phase of a test.