Adds tests for the Avionics actor system
To get around Akka TestKit incompatibilities, we create a scoped trait that requires itself to be mixed into a TestKitBase, which in turn requires an implicit ActorSystem.  The previous version's specs2 akka testkit collisions over time dsls have been resolved in 2.2.0. However, Number to ints get confused in 2.2.0 when used in the time dsl, so to fix that issue time values are encoded in tests as strings, and then converted to a scala.concurrent.duration.DurationInt, which plays nicely in the time dsls.

Again, because of the limitations caused by Akka and Specs2 collisions, we have to execute within the confines of our ScopedTestKit trait, which is a Scope. Creating the ActorSystem as required, we can get exactly the behavior that we need. Then we write the specifications.

Cakes EventSource to make it a more testable solution.

Adds Altimeter spec