Removes unused implicits and adds a useful String to DurationInt conversion.
The previous version's specs2 akka testkit collisions over time dsls have been resolved in 2.2.0. However, Number to ints get confused in 2.2.0 when used in the time dsl, so to fix that issue time values are encoded in tests as strings, and then converted to a scala.concurrent.duration.DurationInt, which plays nicely in the time dsls.