Refactors PilotBehavior to be PlaneProviderComponent.
Removes Receive from PilotBehavior, moving it into the Pilot actor.
Renames PilotBehavior to be PlaneProviderComponent.
Creates PilotPlaneProvider for Production plane access.
Defines PilotPlaneProvider in terms of the appropriate TestActorRef for testing.

Changing the PilotBehavior interface makes it easy to test the tightly integrated plane children in the context of the tests.